Adding first draft version.
This version is quite rough but it makes a "simple" job and is functional.
Improvements must be made, specifically for the testing of the asking.
A linear mode will be setup (linear in opposition to random) which will allow
to inject answers and check the mecanism to run through the different items.
And the reader for the AskQuestions function must not be os.Stdin but a reader.
So, definitely, improvements must be done :-)