Adding linear mode for questions.
There are 2 modes available: one is linear (always the same order...)
and another one is random.
The user can choose with the -m option (-m linear). Default is random.
The linear mode was introduced with adding a potential other read for the
stream that interact with the user. The goal is, at last, to check the order
of questions and answers. With random, this would not be possible. But linear
mode makes this possible.