Fixing bug of display and improving formatting.
There was a bug that was destroying display. I was pushing empty lines
to the channel instead of dropping them.
This is the role of the "decorator" thing to decide if we have to jump
lines or not. So no empty lines or items cannot be published to the
qachan.
The decoration thing had to be improved to come back to the original
that was done in my previous version of the program.